Replace exec with debconf resources on Debian
This commit replaces the currently used exec resources on Debian.
These exec resource are used to update the debconf database with
the selected timezone settings. They use programs provided by the
debconf-utils package, which is an optional package in Debian.
This causes failures if the package is not installed.

The debconf type is a native type that only requires the debconf
package. This package should be installed on all Debian based
systems since it is a required package.