Fix /etc/localtime to be an actual symlink
With the existing code, the /etc/localtime file was created as a hard
link but what is needed by timedatectl is a symlink.

With this patch, /etc/localtime will be created as a symbolic link to
point to /usr/share/zoneinfo/$timezone.
It was tested in a real environment and I was able to configure the
timezone to Japan area, it worked well.