Fix link management for localtime
In puppet in order to manage a link, it should use target and not
source. Additionally if the previous instance of the file is not a link
we need force set to true to replace it.  This change should ensure that
on the next run, the /etc/localtime is correctly configured as link to
the appropriate timezone file.